Move slack run commands from common to run
