+++ common/controllers/slack.js
@@ -1,71 +1,10 @@
-const commandsFromRun = require('../../run/services/slack/commands');
-const { getAppStatusFromScalingo } = require('../../run/services/slack/app-status-from-scalingo');
 const shortcuts = require('../services/slack/shortcuts');
 const viewSubmissions = require('../services/slack/view-submissions');
 const { environments, deploy, publish } = require('../services/releases');
 const github = require('../services/github');
 const postSlackMessage = require('../services/slack/surfaces/messages/post-message');
-const sendSlackBlockMessage = require('../services/slack/surfaces/messages/block-message');
-
-function _getDeployStartedMessage(release, appName) {
-  return `Commande de déploiement de la release "${release}" pour ${appName} en production bien reçue.`;
-}
 
 module.exports = {
-
-  createAndDeployPixSiteRelease(request) {
-    const payload = request.pre.payload;
-    commandsFromRun.createAndDeployPixSiteRelease(payload);
-
-    return {
-      'text': _getDeployStartedMessage(payload.text, 'PIX site and pro')
-    };
-  },
-
-  createAndDeployPixUIRelease(request) {
-    const payload = request.pre.payload;
-    commandsFromRun.createAndDeployPixUI(payload);
-
-    return {
-      'text': _getDeployStartedMessage(payload.text, 'PIX UI')
-    };
-  },
-
-  createAndDeployEmberTestingLibraryRelease(request) {
-    const payload = request.pre.payload;
-    commandsFromRun.createAndDeployEmberTestingLibrary(payload);
-
-    return {
-      'text': _getDeployStartedMessage(payload.text, 'EMBER-TESTING-LIBRARY')
-    };
-  },
-
-  createAndDeployPixBotRelease(request) {
-    const payload = request.pre.payload;
-    commandsFromRun.createAndDeployPixBotRelease(payload);
-
-    return {
-      'text': _getDeployStartedMessage(payload.text, 'PIX Bot')
-    };
-  },
-
-  createAndDeployPixLCMSRelease(request) {
-    const payload = request.pre.payload;
-    commandsFromRun.createAndDeployPixLCMS(payload);
-
-    return {
-      'text': _getDeployStartedMessage(payload.text, 'PIX LCMS')
-    };
-  },
-
-  createAndDeployPixDatawarehouseRelease(request) {
-    const payload = request.pre.payload;
-    commandsFromRun.createAndDeployPixDatawarehouse(payload);
-
-    return {
-      'text': _getDeployStartedMessage(payload.text, 'PIX Datawarehouse')
-    };
-  },
 
   createAndDeployPixHotfix(request) {
     const payload = request.pre.payload;
@@ -78,23 +17,6 @@ module.exports = {
     return {
       'text': 'Commande de déploiement de hotfix de PIX en recette.'
     };
-  },
-
-  getAppStatus(request) {
-    const appName= request.pre.payload.text;
-    return getAppStatusFromScalingo(appName);
-  },
-
-  async deployLastVersion(request) {
-    const appName = request.pre.payload.text;
-
-    try {
-      await commandsFromRun.getAndDeployLastVersion({ appName });
-    } catch(e) {
-      return sendSlackBlockMessage(e.message);
-    }
-
-    return sendSlackBlockMessage(`Re-déploiement de ${appName} déclenché`);
   },
 
   interactiveEndpoint(request) {
